# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Airpak Express app API — shipments, tracking, messages, payments, admin simulation.
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthSession {
  user: User;
}