# lib/api-zod/src/generated/types/messageThread.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Airpak Express app API — shipments, tracking, messages, payments, admin simulation.
 * OpenAPI spec version: 0.1.0
 */

export interface MessageThread {
  id: number;
  subject: string;
  lastMessage: string;
  lastMessageAt: Date;
  unreadCount: number;
  participantName: string;
  /** @nullable */
  shipmentId?: number | null;
}